# Add `perf` subcommand to benchmark Adroit compiler