Previously e.g. in #60 and #61 I would manually insert some code into `crates/adroit/src/compile.rs` to measure and print timing data, then manually gather the data and plug all the numbers into a calculator. This PR gets rid of the need for all that by adding an `adroit perf` subcommand that you can use like this:

```sh
cargo run --release perf -n100000 evals/gmm/gmm.adroit
```

```
100000 × lex       =  0.869 seconds
100000 × parse     =  2.675 seconds
100000 × typecheck = 25.281 seconds
100000 × total     = 28.825 seconds

evals/gmm/gmm.adroit is 139 lines
lines of code per second: 482218
```